+++ src/components/Login/Login.tsx
@@ -5,14 +5,14 @@ import { InputWindow } from '../InputSection/InputWindow';
 import { signinUser } from '@/api/userAuth';
 import { useRouter } from 'next/navigation';
 import { HideToggle } from '../Toggle/HideToggle';
-import { useStore } from '@/store/useAuthStore';
+import { useAuthStore } from '@/store/useAuthStore';
 import { getUserData } from '@/api/getUserData';
 
 const Login = () => {
 	const router = useRouter();
 	const [referrer, setReferrer] = useState<string | null>(null); // referrer 상태 추가
 	const { isLoggedIn, token, userId, setIsLoggedIn, setToken, setUserId } =
-		useStore(); //zustand 상태
+		useAuthStore(); //zustand 상태
 	const debouncingTimer = useRef<NodeJS.Timeout | null>(null);
 	const [id, setId] = useState('');
 	const [password, setPassword] = useState('');
